fix: Missing fields on protocol parameters
